// src/app/types/types.ts
export interface BusinessCard {
  id: string;
  name: string;
  title: string;
  company: string;
  email: string;
  phone: string;
  website: string;
  address: string;
  description: string;
  avatarImage: string;
  backgroundImage: string;
  userId: string;
  createdAt: string;
  updatedAt: string;
  language: string;
}